Fixed Swerve! Intake Code?
Swerve still has a glitch, but the offsets, negatives/positives, and ports have been updated. In addition, I added some basic intake code (which may or may not work) which is commented out.
Lastly, I added a try catch loop on the instantiation of every subsystem so on test bots we can easily test one at a time.