Made Auto Swerve Accurate
I fixed the comments on which direction is which, added an integral value to the control to get more accurate, and increased the speed to over double the previous speed. This gets to within 1cm of accuracy.
However, the robot is spinning while swerving currently, which is not good... We need to fix this; I'll add it as an issue.