Overhauled Shooter, PathPlanner, and Shuffleboard.
Essentially, the shooter now has the code to allow it to rotate to change the angle we're shooting from (it may or may not work as of now due to tuning PID, Feedforward, and positive/negative issues).
PathPlanner has been for the most part fixed, but rotation while moving still does not work.
The Shuffleboard is starting to look nice as well.